+++ src/Cogworks.AzureSearch.IoC.Autofac/Cogworks.AzureSearch.IoC.Autofac.csproj
