+++ src/Cogworks.AzureSearch.IoC.Autofac/Extensions/AutofacExtensions.cs
@@ -1,7 +1,7 @@
-﻿using AutofacContainerBuilder = Autofac.ContainerBuilder;
-using Cogworks.AzureSearch.Autofac.Builders;
+﻿using Cogworks.AzureSearch.IoC.Autofac.Builders;
+using AutofacContainerBuilder = Autofac.ContainerBuilder;
 
-namespace Cogworks.AzureSearch.Autofac.Extensions
+namespace Cogworks.AzureSearch.IoC.Autofac.Extensions
 {
     public static class AutofacExtensions
     {
