+++ src/Cogworks.AzureSearch.IoC.LightInject/Cogworks.AzureSearch.IoC.LightInject.csproj
@@ -13,7 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <InternalsVisibleTo Include="Cogworks.AzureSearch.Umbraco.IocExtension" />
+    <InternalsVisibleTo Include="Cogworks.AzureSearch.IoC.Umbraco" />
   </ItemGroup>
 
   <ItemGroup>
