+++ src/Cogworks.AzureSearch.IoC.LightInject/Cogworks.AzureSearch.IoC.LightInject.nuspec
