+++ src/Cogworks.AzureSearch.IoC.Microsoft/Cogworks.AzureSearch.IoC.Microsoft.csproj
