+++ src/Cogworks.AzureSearch.IoC.Umbraco/Cogworks.AzureSearch.IoC.Umbraco.nuspec
