+++ src/Cogworks.AzureSearch.IoC.Umbraco/app.config
