+++ src/Cogworks.AzureSearch.IoC.Umbraco/packages.config
