+++ src/Cogworks.AzureSearch/Cogworks.AzureSearch.csproj
@@ -14,10 +14,10 @@
 
   <!-- Project Visibility -->
   <ItemGroup>
-    <InternalsVisibleTo Include="Cogworks.AzureSearch.Autofac.IocExtension" />
-    <InternalsVisibleTo Include="Cogworks.AzureSearch.Umbraco.IocExtension" />
-    <InternalsVisibleTo Include="Cogworks.AzureSearch.LightInject.IocExtension" />
-    <InternalsVisibleTo Include="Cogworks.AzureSearch.Microsoft.IocExtension" />
+    <InternalsVisibleTo Include="Cogworks.AzureSearch.IoC.Autofac" />
+    <InternalsVisibleTo Include="Cogworks.AzureSearch.IoC.Umbraco" />
+    <InternalsVisibleTo Include="Cogworks.AzureSearch.IoC.LightInject" />
+    <InternalsVisibleTo Include="Cogworks.AzureSearch.IoC.Microsoft" />
   </ItemGroup>
 
   <!-- Tests Visibility -->
