+++ tests/UnitTests/Cogworks.AzureSearch.LightInject.UnitTests/Cogworks.AzureSearch.LightInject.UnitTests.csproj
@@ -35,6 +35,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\Cogworks.AzureSearch.LightInject.IocExtension\Cogworks.AzureSearch.LightInject.IocExtension.csproj" />
+    <ProjectReference Include="..\..\..\src\Cogworks.AzureSearch.IoC.LightInject\Cogworks.AzureSearch.IoC.LightInject.csproj" />
   </ItemGroup>
 </Project>
